refactor: api config
